style(configs): rename defaultConfigs to configs
